Make credstash honor ~/.aws/config region setting
Currently credstash manually replicate parts of Boto3's configuration data
parsing - it reads its command line argument, falling back to
AWS_DEFAULT_REGION, and after that falling back to us-east-1.

Per the Boto3 documentation:
https://boto3.readthedocs.org/en/latest/guide/configuration.html

it should also read ~/.aws/config for the region setting if it is not set
in the environment.

To avoid reimplementing the configuration parsing in achieving this:
- use the args.region command line argument (if unset this will be None,
  which is the default for Boto3 methods anyway).
- try and get the dynamodb resource using this region setting
- if it fails, then as a last resort use the DEFAULT_REGION setting
  hard-coded to us-east-1