Fix the test for whether a region is valid.
We weren't actually checking whether the region passed on the
command-line was valid, but rather whether there was a default region
available via either .aws/config or the environment. If there was no
such region, we would unconditionally fall back on DEFAULT_REGION,
instead of trying the one passed on the command line.