Allow default version for putSecret
The logic for the putSecret function previously required a version, but performed a default check in the function to add a padded 1 if the version is set to an empty string. This commit sets `version` to an empty string by default in order to take advantage of the default value when a library user does not provide a version. This change should not break current integrations with the library.