Factor out encrypt and decrypt functions.
We are integrating credstash logic into our application stack so we can load runtime
configuration from encrypted DynamoDB storage. Since the credstash tool is designed
primarily as a CLI, we've had to duplicate credstash logic in our application code.

In this PR, I'm attempting to factor out the `encrypt` and `decrypt` operations so
we can use them directly with credstash as a library.