Move --key parameter to super parameters
Means it's available to all commands (previously it was possible to specify in `put` but not in `get`).

This is a breaking change in the sense that you used to have to write `credstash put keyname value --key='alias/somekms'` and now you have to write `credstash --key='alias/somekms' put keyname value`.  But since it was impossible to use any of the `get` commands with non-standard KMS, I conclude that no one can be using the deprecated syntax!