Set default EncryptionContext at KMS
Using the credential name and version as Additional Authentication Data (AAD)
can mitigate ciphertext replacement attacks on the DynamoDB backend.

Bumps the version to 2.0.0 as it is backwards incompatible with older versions
of credstash which didn't set an EncryptionContext by default.

Upgrading from older versions is possible by running

  ./migrations/credstash-migrate-encryption-context.py

This will re-encrypt secrets with the default EncryptionContext.