Resolves #73: Adds binary secret support
This is perhaps a superficial approach to supporting binary secrets. If
it's not satisfactory, I'd love to discuss what would be, so I can help
get that on the road. I shamelessly ripped the binary check function
from jfs of SO: https://stackoverflow.com/a/7392391/830495.
I didn't want to introduce any external dependencies to add this
support, but maybe there's a better option? I'm unsure.